requirements.txt
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
coverage==4.5.4
dateutils==0.6.8
distlib==0.3.1
docutils==0.16
filelock==3.0.12
friendlypins==1.0.4
idna==2.10
imagesize==1.2.0
iniconfig==1.0.1
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
macholib==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
packaging==20.4
pbr==5.4.5
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
Pygments==2.6.1
PyInstaller==3.6
pylint==2.5.3
pyparsing==2.4.7
PySide2==5.15.0
pytest==6.0.1
pytest-cov==2.10.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
QtPy==1.9.0
requests==2.24.0
shiboken2==5.15.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinx-qt-documentation==0.3
sphinx-rtd-theme==0.5.0
sphinxcontrib-apidoc==0.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.1
tox==3.18.1
tox-factor==0.1.2
tox-pyenv==1.1.0
tqdm==4.48.0
urllib3==1.25.10
virtualenv==20.0.29
wrapt==1.12.1
